+++ packages/terra-application-navigation/src/tabs/_Tabs.jsx
@@ -9,6 +9,7 @@ import Popup from 'terra-popup';
 import Tab from './_Tab';
 import TabRollup from './_TabRollup';
 import PopupMenu from '../common/_PopupMenu';
+import { navigationItemId } from '../utils/helpers';
 
 import {
   navigationItemsPropType,
@@ -23,6 +24,10 @@ const propTypes = {
    * An array of configuration objects with information specifying the creation of navigation items.
    */
   navigationItems: navigationItemsPropType,
+  /**
+   * The base id used to generate ids of navigation, utility, and extension items
+   */
+  id: PropTypes.string,
   /**
    * A function to be executed for the render of each navigation item.
    */
@@ -207,11 +212,12 @@ class Tabs extends React.Component {
     };
   }
 
-  renderVisibleTabs(visibleTabs, useNotificationStyle, notifications) {
+  renderVisibleTabs(visibleTabs, useNotificationStyle, notifications, id) {
     const { activeTabKey, onTabSelect } = this.props;
 
     return visibleTabs.map((tab, index) => {
       const tabProps = {
+        id: id && navigationItemId(id, tab.key),
         text: tab.text,
         key: tab.key,
         onTabSelect: onTabSelect ? onTabSelect.bind(null, tab.key, tab.metaData) : null,
@@ -251,7 +257,7 @@ class Tabs extends React.Component {
     );
   }
 
-  renderPopup(hiddenTabs, notifications) {
+  renderPopup(hiddenTabs, notifications, id) {
     const {
       activeTabKey, onTabSelect, intl,
     } = this.props;
@@ -272,6 +278,7 @@ class Tabs extends React.Component {
           title={intl.formatMessage({ id: 'Terra.applicationNavigation.tabs.rollupMenuHeaderTitle' })}
           role="list"
           menuItems={hiddenTabs.map(tab => ({
+            id: id && navigationItemId(id, tab.key),
             key: tab.key,
             text: tab.text,
             icon: tab.icon,
@@ -296,6 +303,7 @@ class Tabs extends React.Component {
     const {
       navigationItems,
       notifications,
+      id,
     } = this.props;
     if (!navigationItems || !navigationItems.length) {
       return <Tab isPlaceholder text="W" tabKey="" aria-hidden="true" />;
@@ -310,9 +318,9 @@ class Tabs extends React.Component {
         className={cx('tabs-container', { 'is-calculating': this.isCalculating })}
         ref={this.containerRef}
       >
-        {this.renderVisibleTabs(visibleTabs, useNotificationStyle, notifications)}
+        {this.renderVisibleTabs(visibleTabs, useNotificationStyle, notifications, id)}
         {!this.menuHidden ? this.renderRollup(hiddenTabs, useNotificationStyle, notifications) : null}
-        {popupIsOpen ? this.renderPopup(hiddenTabs, notifications) : null}
+        {popupIsOpen ? this.renderPopup(hiddenTabs, notifications, id) : null}
       </nav>
     );
   }